@@ -3,24 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@fortawesome/fontawesome-free": "5.15.3",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "antd": "^4.16.13",
-    "axios": "^0.24.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "5.2.0",
+    "@ant-design/icons": "^6.0.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@fortawesome/fontawesome-free": "6.7.2",
+    "@mui/material": "^7.1.1",
+    "@mui/styled-engine": "^7.1.1",
+    "@mui/system": "^7.1.1",
+    "@mui/x-date-pickers": "^8.5.2",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "antd": "^5.26.1",
+    "axios": "^1.10.0",
+    "dayjs": "^1.11.13",
+    "moment": "^2.30.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-redux": "^9.2.0",
+    "react-router": "^7.6.2",
+    "react-router-dom": "^7.6.2",
     "react-scripts": "^5.0.1",
-    "redux": "^4.1.2",
+    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.4.0",
-    "web-vitals": "^1.1.2",
-    "webpack": "^5.97.1"
+    "redux-thunk": "^3.1.0",
+    "web-vitals": "^5.0.3",
+    "webpack": "^5.99.9"
   },
   "scripts": {
     "start": "set PORT=3006 && react-scripts start",
@@ -46,5 +55,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "css-loader": "^7.1.2"
   }
 }
