@@ -1,5 +1,4 @@
-import { getConfig } from "@testing-library/react";
-
+// In src/helpers/helper.js (or src/helpers.js, depending on your file name)
 export const Helpers = {
   setCookie(cName, cValue, expDays) {
     let date, expires;
@@ -13,9 +12,10 @@ export const Helpers = {
       ? cName + "=" + cValue + "; " + expires + "; path=/"
       : cName + "=" + cValue + "; path=/";
   },
+
   getCookie(cName) {
     const name = cName + "=";
-    const cDecoded = decodeURIComponent(document.cookie); //to be careful
+    const cDecoded = decodeURIComponent(document.cookie);
     const cArr = cDecoded.split("; ");
     let res;
     cArr.forEach((val) => {
@@ -23,44 +23,49 @@ export const Helpers = {
     });
     return res;
   },
-  deleteAllCookies() {
-    var cookies = document.cookie.split(";");
 
-    for (var i = 0; i < cookies.length; i++) {
-      var cookie = cookies[i];
-      var eqPos = cookie.indexOf("=");
-      var name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
+  deleteAllCookies() {
+    const cookies = document.cookie.split(";");
+    for (let i = 0; i < cookies.length; i++) {
+      const cookie = cookies[i];
+      const eqPos = cookie.indexOf("=");
+      const name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie;
       document.cookie = name + "=;expires=Thu, 01 Jan 1970 00:00:00 GMT";
     }
   },
+
   getResourceUrl(url) {
-    return `${getConfig().API_BASE_URL}/${url}`;
+    const API_BASE_URL = process.env.REACT_APP_API_BASE_URL || 'http://localhost:8000';
+    return `${API_BASE_URL}/${url.replace(/^\//, '')}`;
   },
+
   percentageValidator(fieldName) {
     return () => ({
       validator(_, value) {
-        if (value && (value > 100 || value < 1)) {
-          return Promise.reject(`${fieldName} must be in between 1-100`);
+        if (value === undefined || value === null || value < 0 || value > 100) {
+          return Promise.reject(new Error(`${fieldName} must be between 0 and 100`));
         }
         return Promise.resolve();
       },
     });
   },
+
   positiveValidator(fieldName) {
     return () => ({
       validator(_, value) {
-        if (value && value < 1) {
-          return Promise.reject(`${fieldName} must be greater than 0`);
+        if (value === undefined || value === null || value <= 0) {
+          return Promise.reject(new Error(`${fieldName} must be greater than 0`));
         }
         return Promise.resolve();
       },
     });
   },
-  morethanZeroValidator(fieldName) {
+
+  moreThanZeroValidator(fieldName) {
     return () => ({
       validator(_, value) {
-        if (value && value < 0) {
-          return Promise.reject(`${fieldName} must be greater than 0`);
+        if (value === undefined || value === null || value <= 0) {
+          return Promise.reject(new Error(`${fieldName} must be greater than 0`));
         }
         return Promise.resolve();
       },
