@@ -2,27 +2,25 @@ import { Dashboard } from "./containers/dashboard";
 import { PredictionForm } from "./containers/prediction-form";
 import { ReadyTempaltes } from "./containers/templates";
 
-var routes = [
+const routes = [
   {
     path: "/dashboard",
     name: "Dashboard",
     icon: "ni ni-books text-primary",
-    component: Dashboard,
-    exact: true
+    element: <Dashboard />,
   },
   {
     path: "/ev-analysis",
     name: "EV Analysis",
     icon: "ni ni-email-83 text-primary",
-    component: PredictionForm,
-    exact: true
+    element: <PredictionForm />,
   },
   {
     path: "/templates",
     name: "Templates",
     icon: "ni ni-email-83 text-primary",
-    component: ReadyTempaltes,
-    exact: true
+    element: <ReadyTempaltes />,
   },
 ];
+
 export default routes;
